+++ src/core/ReflMngr.cpp
@@ -794,7 +794,7 @@ Name ReflMngr::AddDestructor(Type type) {
 					Mngr.Destruct(ObjectView{ basetype, baseptr });
 				}
 			},
-			MethodFlag::Const
+			MethodFlag::Variable
 		} }
 	);
 }
@@ -1253,7 +1253,7 @@ bool ReflMngr::IsDestructible(Type type) const {
 	if (begin_iter == end_iter)
 		return true;
 	for (auto iter = begin_iter; iter != end_iter; ++iter) {
-		if (iter->second.methodptr.GetMethodFlag() == MethodFlag::Const
+		if (iter->second.methodptr.GetMethodFlag() == MethodFlag::Variable
 			&& IsCompatible(iter->second.methodptr.GetParamList(), {}))
 			return true;
 	}
@@ -1292,7 +1292,7 @@ bool ReflMngr::Destruct(ObjectView obj) const {
 		return true;// trivial ctor
 	auto [begin_iter, end_iter] = typeinfo.methodinfos.equal_range(NameIDRegistry::Meta::dtor);
 	for (auto iter = begin_iter; iter != end_iter; ++iter) {
-		if (iter->second.methodptr.GetMethodFlag() == MethodFlag::Const
+		if (iter->second.methodptr.GetMethodFlag() == MethodFlag::Variable
 			&& IsCompatible(iter->second.methodptr.GetParamList(), {}))
 		{
 			iter->second.methodptr.Invoke(obj.GetPtr(), nullptr, {});